feat(plugins): Configuration stage 4a
