+++ crates/facet_generate/src/generation/csharp/installer/runtime/core/Unit.cs
@@ -1,0 +1,3 @@
+namespace Facet.Runtime.Serde;
+
+public readonly record struct Unit;
