+++ crates/facet_generate/src/generation/csharp/installer/runtime/serde/DeserializationError.cs
@@ -1,0 +1,10 @@
+using System;
+
+namespace Facet.Runtime.Serde;
+
+public sealed class DeserializationError : Exception
+{
+    public DeserializationError(string message) : base(message)
+    {
+    }
+}
