+++ crates/facet_generate/src/generation/csharp/installer/runtime/serde/SerializationError.cs
@@ -1,0 +1,10 @@
+using System;
+
+namespace Facet.Runtime.Serde;
+
+public sealed class SerializationError : Exception
+{
+    public SerializationError(string message) : base(message)
+    {
+    }
+}
